backend/requirements.txt
fastapi
uvicorn
sqlite-bro
sqlparse
python-dotenv
groq
numpy
sentence-transformers
pandas
